refactor(greeter): consolidate plugin into plugins/greeter/
- Move GreeterPanel component, QuizMode, PanelSettingsModalGreeter, and
  configModal into plugins/greeter/; merge configModal + settingsModal
  into a single configModal.tsx
- Move GreeterConfig type into plugins/greeter/types.ts (out of app/types.ts)
- Move useGreeterConfig + GreeterConfigProvider into plugins/greeter/useGreeterConfig.ts
  (out of app/context/PanelConfigs.tsx)
- Move stories and test into plugin dir; update Storybook + vitest configs
  to discover from plugins/**
- Document file naming convention in plugins/README.md: camelCase for
  plugin API contract files, PascalCase for internal components
