+++ app/components/apps/ReactionCanvasAppV4.tsx
@@ -3,11 +3,13 @@ import Canvas from "../shared/Canvas";
 import TouchLayer from "../shared/TouchLayer";
 import AdminPanelNoDB from "../panels/AdminPanelNoDB";
 import InterfaceChipBar from "../shared/InterfaceChipBar";
-import type { ActivityMode, GreeterConfig, SocialConfig, ValenceInputMode } from "../../types";
+import type { ActivityMode, SocialConfig, ValenceInputMode } from "../../types";
 import { PANEL_REGISTRY, SOLO_SCREEN_LABEL } from "../../panelRegistry";
 import type { PanelDefinition } from "../../panelRegistry";
 import { PanelContextProvider } from "../../context/PanelContext";
-import { GreeterConfigProvider, SocialMediaConfigProvider, ImageCanvasConfigProvider } from "../../context/PanelConfigs";
+import { GreeterConfigProvider } from "../../../plugins/greeter/useGreeterConfig";
+import type { GreeterConfig } from "../../../plugins/greeter/types";
+import { SocialMediaConfigProvider, ImageCanvasConfigProvider } from "../../context/PanelConfigs";
 import GithubUsernameModal from "../modals/GithubUsernameModal";
 import FeedbackStarsModal from "../modals/FeedbackStarsModal";
 import InterfacePushModal from "../modals/InterfacePushModal";
