+++ plugins/greeter/QuizMode.tsx
