+++ plugins/greeter/server.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from 'vitest';
-import { GreeterServerPlugin } from '../plugins/greeter/server';
-import type { PluginConnection, PluginContext } from '../plugins/types';
+import { GreeterServerPlugin } from './server';
+import type { PluginConnection, PluginContext } from '../types';
 
 function makeCtx(): PluginContext {
   return {
