+++ plugins/greeter/types.ts
@@ -1,1 +1,5 @@
-export type GreeterPluginState = { config: { eventUrl: string } | null };
+export interface GreeterConfig {
+  eventUrl: string;
+}
+
+export type GreeterPluginState = { config: GreeterConfig | null };
